+++ src/uniqat/__init__.py
@@ -45,7 +45,7 @@ from .core.metrics import UnderwaterMetrics
 from .core.video_assessor import VideoQualityAssessment, VideoQualityAssessor
 from .utils.visualization import QualityVisualizer
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 __author__ = "Alzayat Saleh"